adds conditional to update and destroy method so users can only edit and delete their posts, installs jb gem and generates posts index and show

+++ app/controllers/posts_controller.rb
@@ -1,5 +1,12 @@
 class PostsController < ApplicationController
-  before_action :authenticate_user, except: [:index] # makes it so users must be logged in to use these actions, BESIDES the index
+  before_action :authenticate_user, except: [:index] # makes it so users must be logged in to use these actions, excluding index
+
+  # before_action :authenticate_post, except: [:index, :create]
+
+  # def correct_user
+  #   post = Post.find_by(id: params[:id])
+  #   redirect_to unless current_user.id == post.user_id
+  # end
 
   def index
     posts = Post.all
@@ -27,21 +34,29 @@ class PostsController < ApplicationController
   end
 
   def update
-    post = Post.find(params[:id])
-    post.title = params[:title] || post.title
-    post.body = params[:body] || post.body
-    post.user_id = current_user.id
-    post.game_id = params[:game_id]
-    if post.save
-      render json: post
+    post = Post.find_by(id: params[:id])
+    if current_user.id == post.user_id
+      post.title = params[:title] || post.title
+      post.body = params[:body] || post.body
+      post.user_id = current_user.id
+      post.game_id = params[:game_id]
+      if post.save
+        render json: post
+      else
+        render json: { errors: post.errors.full_messages }, status: :unprocessable_entity
+      end
     else
-      render json: { errors: post.errors.full_messages }, status: :unprocessable_entity
+      render json: {}, status: :unauthorized
     end
   end
 
   def destroy
     post = Post.find_by(id: params[:id])
-    post.destroy
-    render json: { message: "Post successfully deleted." }
+    if current_user.id == post.user_id
+      post.destroy
+      render json: { message: "Post successfully deleted." }
+    else
+      render json: {}, status: :unauthorized
+    end
   end
 end
